+++ test/extensions/common/aws/aws_metadata_fetcher_integration_test.cc
@@ -25,7 +25,7 @@ public:
           "@type": type.googleapis.com/envoy.config.filter.network.http_connection_manager.v2.HttpConnectionManager
           stat_prefix: metadata_test
           http_filters:
-            - name: envoy.fault
+            - name: envoy.filters.http.fault
               typed_config:
                 "@type": type.googleapis.com/envoy.config.filter.http.fault.v2.HTTPFault
                 delay:
@@ -35,7 +35,7 @@ public:
                   percentage:
                     numerator: 100
                     denominator: HUNDRED
-            - name: envoy.router
+            - name: envoy.filters.http.router
           codec_type: HTTP1
           route_config:
             virtual_hosts: